refactor text_preprocessing.py and vectorization.py
