+++ src/text_preprocessing.py
@@ -1,11 +1,9 @@
-import nltk
-nltk.download('stopwords')
-from nltk.corpus import stopwords
-from ast import literal_eval
-import pandas as pd
-import numpy as np
 import re
+from ast import literal_eval
 
+import joblib
+import pandas as pd
+from nltk.corpus import stopwords
 
 
 def read_data(filename):
@@ -13,11 +11,13 @@ def read_data(filename):
     data['tags'] = data['tags'].apply(literal_eval)
     return data
 
+
 def split_data(train, validation, test):
     X_train, y_train = train['title'].values, train['tags'].values
     X_val, y_val = validation['title'].values, validation['tags'].values
     X_test = test['title'].values
     return X_train, y_train, X_val, y_val, X_test
+
 
 def text_prepare(text):
     """
@@ -28,38 +28,38 @@ def text_prepare(text):
     REPLACE_BY_SPACE_RE = re.compile('[/(){}\[\]\|@,;]')
     BAD_SYMBOLS_RE = re.compile('[^0-9a-z #+_]')
     STOPWORDS = set(stopwords.words('english'))
-    text = text.lower() # lowercase text
-    text = re.sub(REPLACE_BY_SPACE_RE, " ", text) # replace REPLACE_BY_SPACE_RE symbols by space in text
-    text = re.sub(BAD_SYMBOLS_RE, "", text) # delete symbols which are in BAD_SYMBOLS_RE from text
-    text = " ".join([word for word in text.split() if not word in STOPWORDS]) # delete stopwords from text
+    text = text.lower()  # lowercase text
+    text = re.sub(REPLACE_BY_SPACE_RE, " ", text)  # replace REPLACE_BY_SPACE_RE symbols by space in text
+    text = re.sub(BAD_SYMBOLS_RE, "", text)  # delete symbols which are in BAD_SYMBOLS_RE from text
+    text = " ".join([word for word in text.split() if not word in STOPWORDS])  # delete stopwords from text
     return text
-
-def test_text_prepare():
-    examples = ["SQL Server - any equivalent of Excel's CHOOSE function?",
-                "How to free c++ memory vector<int> * arr?"]
-    answers = ["sql server equivalent excels choose function", 
-               "free c++ memory vectorint arr"]
-    for ex, ans in zip(examples, answers):
-        if text_prepare(ex) != ans:
-            return "Wrong answer for the case: '%s'" % ex
-    return 'Basic tests are passed.'
 
 
 def main():
-    train = read_data('data/train.tsv')
-    validation = read_data('data/validation.tsv')
-    test = pd.read_csv('data/test.tsv', sep='\t')
+    data_directory = "../data"
+
+    # Read data
+    train = read_data(data_directory + '/train.tsv')
+    validation = read_data(data_directory + '/validation.tsv')
+    test = pd.read_csv(data_directory + '/test.tsv', sep='\t')
+
+    # Split data
     X_train, y_train, X_val, y_val, X_test = split_data(train, validation, test)
+
+    # Prepare tests
     prepared_questions = []
-    for line in open('data/text_prepare_tests.tsv', encoding='utf-8'):
+    for line in open(data_directory + '/text_prepare_tests.tsv', encoding='utf-8'):
         line = text_prepare(line.strip())
         prepared_questions.append(line)
+
     text_prepare_results = '\n'.join(prepared_questions)
     X_train = [text_prepare(x) for x in X_train]
     X_val = [text_prepare(x) for x in X_val]
     X_test = [text_prepare(x) for x in X_test]
+
     # Dictionary of all tags from train corpus with their counts.
     tags_counts = {}
+
     # Dictionary of all words from train corpus with their counts.
     words_counts = {}
 
@@ -76,13 +76,17 @@ def main():
                 tags_counts[tag] += 1
             else:
                 tags_counts[tag] = 1
-            
+
     # print(tags_counts)
     # print(words_counts)
+    #
+    # print(sorted(words_counts, key=words_counts.get, reverse=True)[:3])
+    # most_common_tags = sorted(tags_counts.items(), key=lambda x: x[1], reverse=True)[:3]
+    # most_common_words = sorted(words_counts.items(), key=lambda x: x[1], reverse=True)[:3]
 
-    print(sorted(words_counts, key=words_counts.get, reverse=True)[:3])
-    most_common_tags = sorted(tags_counts.items(), key=lambda x: x[1], reverse=True)[:3]
-    most_common_words = sorted(words_counts.items(), key=lambda x: x[1], reverse=True)[:3]
+    joblib.dump((X_train, X_val, X_test), "../output/X_preprocessed.joblib")
+    joblib.dump((y_train, y_val), "../output/y_preprocessed.joblib")
+    joblib.dump(words_counts, "../output/words_counts.joblib")
 
 
 if __name__ == "__main__":
